+++ mod/primitives/pkg/ssz/merkleize.go
@@ -247,6 +247,7 @@ func Merkleize[U64T U64[U64T], RootT ~[32]byte](
 		return effectiveChunks[0], nil
 	}
 
+	// TODO: reuse the same hasher for reduced memory allocations.
 	hasher := merkle.NewHasher[RootT]()
 	return hasher.NewRootWithMaxLeaves(
 		effectiveChunks,
